[prices] init: Add strict mode configuration
Strict mode helps developers detekt development issues early and forces
them to fix those. Otherwise, the debug version of the app will
complain at runtime on specific thread or vm policy violation.

The most relevant example of how strict mode helps develop more robust
apps, is when it complains about disk reads/writes violations that
happen on the UI thread, forcing the developer to think twice about
doing that or change that to a background thread in order to unblock
the UI thread.