[prices] init: Add view model dependency injection
The ViewModelFactory will be used with the BaseActivity, which will
hold the responsibility of instantiating the correct ViewModel for the
specific activity, through Dagger sub-modules.

This change enhances the BaseActivity so that it knows how to construct
an architecture components ViewModel specific to the contract of the
activity. Everything is connected with Dagger sub-modules, which makes
sure that the Activity gets the correct ViewModel bind to it.